requirements.txt
flask
pillow
numpypip
opencv-python
azure-cognitiveservices-vision-computervision
python-dotenv